requirements.txt
ipykernel
openpyxl
numpy
pandas
matplotlib
seaborn
scikit-learn
imblearn
statsmodels
flask
gunicorn
mysql-connector-python